+++ src/common/Nav/Results/index.js
@@ -1,0 +1,5 @@
+
+
+export const Results = () => {
+    <p>test</p>
+};
