+++ src/common/Nav/index.js
@@ -4,21 +4,15 @@ import {
   Title,
   Video,
   List,
-  Search,
-  Magnifier,
-  InputStyled,
   StyledLink,
   StyledNavLink,
   Section,
   ContentWidth
 } from "./styled";
 import video from "../../images/Video.svg";
-import magnifier from "../../images/search.svg";
-import { useLocation } from "react-router-dom";
+import { Search } from "./SearchInput";
 
 export const Nav = () => {
-  const location = useLocation();
-
   return(
   <Wrapper>
     <ContentWidth>
@@ -38,17 +32,7 @@ export const Nav = () => {
           </li>
         </List>
       </Section>
-        <Search>
-          <Magnifier src={magnifier} />
-          <InputStyled 
-            placeholder={
-              `Search for
-              ${location.pathname.includes("movies" 
-                      ? "movies..." 
-                      : "people...")
-              }`} 
-          />
-        </Search>
+      <Search/>
     </ContentWidth>
   </Wrapper>
   );
